refactor(userlist): move UserListsViewModel and Factory to userlist.presentation.viewmodel
